move database code to a service folder
